requirements.txt
discord.py
python-decouple
youtube_dl
PyNaCL
ffmpeg